float

+++ python_introduction/finance_calculator.py
@@ -1,6 +1,6 @@
 income = int(input("Enter your monthly income: "))
 expense = int(input("Enter your total monthly expenses: "))
-monthly_savings = int(income - expense)
-savings = int(monthly_savings * 12 + (monthly_savings * 12 * 0.05))
+monthly_savings = income - expense
+savings = monthly_savings * 12 + (monthly_savings * 12 * 0.05)
 print("Your montly savings are $", monthly_savings)
 print("Projected savings after one year, with intrester, is:", savings)
